Handle corruption for interrupted pack
The most important objective is to prevent the data slab and hashes slab from
getting corrupted and loosing archived data.  Incomplete writes during a pack to
the slabs should be the only way for the slabs to get in an inconsistent state.
To allow us to detect and correct this we introduce a check point file at the
root of the archive which is written and sync'd to stable storage before we
start the pack operation.  This way if the pack operation is interrupted, we
can put the slab files back to where they were before we started with a repair
option to the verify-all sub command.  Moving forward, the idea is we add the
ability to periodically update the checkpoint for long running operations by
quiescing IO to the data slab, hashes slab, offsets files, and the stream
output and recording the offset into the input data.  Then we can resume the
operation by checking the files, truncating where needed, and then resuming the
de-dupe operation.

Note: If the slab file and the hashes file have no corruption and the
number of slabs match between the data and hash slab, the slab files are not
touched!  Thus the archive size could be much larger than would be
indicated by the listing of the archive.

Signed-off-by: Tony Asleson <tasleson@redhat.com>