CDC: Force a segment
If we are processing a byte stream that is mostly a repeating byte
pattern we will fail to find any data to consume.  The CDC algr.
implementation doesn't have an upper bound.  This is simple
change to force a consume length when we haven't found one.

We may want to replace the current CDC implementation with QuickCDC.

Without this change, if you try to pack a large zero filled file
you will panic, e.g.

   thread 'main' panicked at src/hash_index.rs:64:13:
   assertion failed: e.len <= u32::MAX as usize

Signed-off-by: Tony Asleson <tasleson@redhat.com>