[stats] Don't use file sizes for data_written & hashes_written
Instead of utilizing the sizes of the data file and hashes file, we'll
keep a running talley of each while processing the file.  This approach
will continue to work when we introduce a client/server arch. with
multiple clients writing to the same archive concurrently.

Note: The data_written will not be 100% accurate as the data slab is
compressed.  Thus what we write into the slab will be more than what
ends up on disk for the data slab.

Signed-off-by: Tony Asleson <tasleson@redhat.com>