[stream] Use 1 file for all binary stream data
This removes the 1 directory for each stream and
replaces it with 2 slab files which contain all
the stream information. 1 slab contains a mapping
from stream id (u32) to the data that tells us
which data slabs in the second file contain the
meta-data and the stream data.  In any cases where
the compressed stream is < 4MiB this will be a 1-1
maping of slabs.  Any of the streams can use 1 or
more slabs to store the stream data as needed.
Thus far, I've not encountered a stream that is >
4MiB.

We leverage the check point to store information
about these 2 slabs so we can put them back to
a known good sync'd point in time.

The directory structure for an archive looks like
this now.

example/
├── data
│   ├── data
│   │   └── slabs
│   │       └── 000
│   │           └── 000
│   │               ├── file_0000000000
│   │               ├── file_0000000000.offsets
│   │               ├── file_0000000001
│   │               ├── file_0000000001.offsets
│   │               ├── file_0000000002
│   │               ├── file_0000000002.offsets
│   │               ├── file_0000000003
│   │               └── file_0000000003.offsets
│   ├── hashes
│   └── hashes.offsets
├── dm-archive.yaml
├── indexes
│   ├── seen
│   └── seen.offsets
├── recovery.checkpoint
└── streams
    ├── mappings 	<- maps -> slabs in meta-data
    ├── mappings.offsets
    ├── metadata	<- metadata with stream data
    └── metadata.offsets

Performance seems good, except that the flight_check
is taking O(N) time on start and needs to change to
do a very basic check, and switch to a more extensive
check if needed.

Signed-off-by: Tony Asleson <tasleson@redhat.com>